pyproject.toml
[project]
name = "beach-time"
version = "0.1.0"
description = ""
authors = [
    {name = "Breno Ramon",email = "brenoramon55@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-ninja (>=1.4.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-ninja-jwt (>=5.3.7,<6.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "task-commit (>=0.1.18,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipython = "^9.2.0"
black = "^25.1.0"
isort = "^6.0.1"

